requirements.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
Django==4.0.4
django-cors-headers==3.11.0
django-dotenv==1.4.2
djangorestframework==3.13.1
numpy==1.22.3
pandas==1.4.2
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
